Added KaiserAir's 737-800's livery
They only own one 737-800 and this is its current livery. If you didn't know KaiseAir existed, that's OK. I didn't know they existed either even though they are based at Oakland International Airport which is 20, 30 minutes north of my house. I only found out they existed after months of trying to figure out who owned a 737 with a flower on its tail that was parked right next to the fence around the airport the last time I was at Oakland's air museum which is across the street from a lot of the hangers on the airport property.